Add admin CRUD

+++ database/migrations/2023_06_02_133348_create_bids_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('bids', function (Blueprint $table) {
             $table->id();
             $table->foreignId('product_id')->constrained()->cascadeOnDelete();
-            $table->smallInteger('value');
+            $table->bigInteger('value');
             $table->timestamps();
         });
     }
